Refactored a lot of the scraper architecture, added framework for adding councils in other states. renamed bayside -> bayside_vic (to avoid conflicts if/when bayside_nsw gets added) added sydney metro councils to councils.md created base scraper class sorting scrapers by state repalced print() statements with logging
