+++ docs/sydneyCouncils.md
@@ -1,0 +1,30 @@
+Bayside Council https://infoweb.bayside.nsw.gov.au/?committee=1
+Blacktown City Council https://www.blacktown.nsw.gov.au/About-Council/How-we-work/Business-Papers
+Municipality of Burwood 
+Camden Council
+City of Campbelltown
+City of Canada Bay
+City of Canterbury-Bankstown
+Cumberland City Council
+Fairfield City Council
+Georges River Council
+Hornsby Shire
+Municipality of Hunter's Hill
+Inner West Council https://www.innerwest.nsw.gov.au/about/the-council/council-meetings/current-council-meetings
+Ku-ring-gai Council
+Lane Cove Council
+City of Liverpool
+Mosman Council
+North Sydney Council
+Northern Beaches Council
+City of Parramatta
+City of Penrith
+City of Randwick
+City of Ryde
+Municipality of Strathfield
+Sutherland Shire
+City of Sydney
+The Hills Shire
+Waverley Council
+City of Willoughby
+Municipality of Woollahra
