+++ scrapers/__init__.py
@@ -1,20 +1,20 @@
-from .maribyrnong import maribyrnong
-from .merribek import merribek
-from .darebin import darebin
-from .boroondara import boroondara
-from .moonee_valley import moonee_valley
-from .melbourne import melbourne
-from .port_phillip import port_phillip
-from .whitehorse import whitehorse
+# from .vic.maribyrnong import maribyrnong
+# from .vic.merribek import merribek
+# from .vic.darebin import darebin
+# from .vic.boroondara import boroondara
+# from .vic.moonee_valley import moonee_valley
+# from .vic.melbourne import melbourne
+# from .vic.port_phillip import port_phillip
+# from .vic.whitehorse import whitehorse
 
 
-councils = [
-    boroondara,
-    darebin,
-    maribyrnong,
-    melbourne,
-    merribek,
-    moonee_valley,
-    port_phillip,
-    whitehorse,
-]
+# councils = [
+#     boroondara,
+#     darebin,
+#     maribyrnong,
+#     melbourne,
+#     merribek,
+#     moonee_valley,
+#     port_phillip,
+#     whitehorse,
+# ]
